chore: move payload factories
