+++ tests/api/articles.integration.spec.ts
@@ -1,9 +1,11 @@
 import {
   ArticlePayload,
+  prepareArticlePayload,
+} from '@_src/api/factories/article-payload.api.factory';
+import {
   Headers,
   apiEndpoints,
   getAuthorizationHeader,
-  prepareArticlePayload,
 } from '@_src/api/utils/api.util';
 import { expect, test } from '@_src/ui/fixtures/merge.fixture';
 import { APIResponse } from '@playwright/test';
