+++ tests/api/comments.integration.spec.ts
@@ -1,10 +1,12 @@
+import { prepareArticlePayload } from '@_src/api/factories/article-payload.api.factory';
 import {
   CommentPayload,
+  prepareCommentPayload,
+} from '@_src/api/factories/comment-payload.api.factory';
+import {
   Headers,
   apiEndpoints,
   getAuthorizationHeader,
-  prepareArticlePayload,
-  prepareCommentPayload,
 } from '@_src/api/utils/api.util';
 import { expect, test } from '@_src/ui/fixtures/merge.fixture';
 import { APIResponse } from '@playwright/test';
